changed

+++ src/main.rs
@@ -12,7 +12,7 @@ fn main() {
     let rank = world.rank();
     let count = world.size() as usize;
     let repeat : usize = 10;
-    let vec_res : Vec<i32> = (rank*repeat..(rank+1)*repeat).collect();
+    let vec_res : Vec<usize> = (rank*repeat..(rank+1)*repeat).collect();
 
     if rank == 0 {
         let mut a = vec![0; count*repeat];
